requirements.txt
asgiref==3.8.1
Django==4.2.13
django-environ==0.11.2
django-rest-framework==0.1.0
djangorestframework==3.15.2
mysqlclient==2.2.4
numpy==2.0.0
opencv-python==4.10.0.84
pillow==10.4.0
sqlparse==0.5.0
typing_extensions==4.12.2
